Uncrush texture nudge buttons in SurfaceInspector
Texture nudge buttons were previously so small that the icon wasn't even
visible on Linux. All nudge buttons are now left/right (I assume it is obvious
that right means "increase" which corresponds to "up" in the vertical
orientation) and the overridden size hints are now removed, so they get the
default minimum size for a small icon button. Also adjusted the icon colour so
the icons show up in both light and dark themes.