Add initial brightness slider to Light Inspector
The wxWidgets colour chooser on GTK is usable for choosing a basic colour, but
it really sucks for adjusting brightness, requiring you to enter a sub-dialog
to choose a "custom colour" (instead of one from its table of presets) and
involving multiple clicks before you can see the result of the light colour
change.

A brightness slider will make it easier to make small adjustments to brightness
without choosing a completely new colour in a separate dialog. The slider is
currently initialised with a brightness value based on the light entity colour,
but moving the slider has no effect yet.