Improve brightness slider behaviour when original colour is black
A multiplicative brightness increase doesn't work when the original colour is
(0, 0, 0), so in this case we just set a grey value based on the new brightness
value.

Normally black lights would not occur, but with the current widget it is very
easy to set the colour to 0 by dragging all the way to the left, at which point
colour information is lost and a subsequent increase in brightness reverts to
grey, rather than the original colour. This should be improved.