Update Light Inspector colour picker after changing brightness
After the slider drag has finished, update the colour picker to show the new
colour. This is multi-selection aware, and will show a transparent colour if
there is more than one light selected and their colours are inconsistent
(transparency might not be the ideal indication for this, but I can't see a
better way to indicate "inconsistent colours" in the wxColourPickerCtrl
documentation).

Note that according to wxWidgets docs, wxEVT_SCROLL_CHANGED only works on
Windows, but it actually seems to work fine on GTK too. Hopefully it will
behave the same way on Windows.