Adjust brightness of all lights in correct proportion
Instead of setting each light to the brightness defined by the current slider
position (which causes dim lights to jump in brightness when selected along
with a bright light), modify each light's starting colour in proportion to the
adjustment delta.

This means that dimming two or more lights together gives the expected result,
even if some of them are much brighter than others.