Brightness adjustment now supports undo
Start an undo command when the drag begins, and finish it when the mouse is
released (i.e. at the same time the colour picker is updated). This avoids
flooding the undo queue with lots of useless tiny adjustment commands.

The incremental drag event is now wxEVT_SCROLL_THUMBTRACK, since an extra
wxEVT_SLIDER is emitted *after* the final wxEVT_SCROLL_CHANGED event, resulting
in a spurious new undo command being initiated which in turn prevents the
previous command from being undone.