Increase size of Fit Texture spinboxes in SurfaceInspector
The minimum size of 8 times the character width is too small on GTK, since the
spin buttons are apparently included in the size allocation. This was resulting
in the boxes appearing as small text entries with no spin buttons available.

Unlike with wxTextCtrl, setting no minimum size results in a spin box which is
too large on GTK, so instead the fixed minimum size is simply doubled from its
previous value.