Initial "Preserve aspect" toggle button in Surface Inspector
The aim of this button is to allow the Fit Texture function to operate on a
single axis, while adjusting the other axis to preserve the texture aspect
ratio. This will be useful for certain textures like wooden planks which only
need an integer number of images on one axis.

Currently the button toggles and has an icon, but does not affect the texture
fitting behaviour.