Initial implementation of single-axis texture fitting
If the button is toggled, a value of -1 will be passed to the implementing
code, indicating that a particular axis should not be fitted. Currently the
result isn't quite correct: rather than preserving aspect ratio, it is
preserving the *current* texture transformation for the non-fitted axis (i.e.
it doesn't touch the shift/scale values for the non-fitted axis at all).