Aspect-preserving texture fitting now working
Abandoned the original plan to calculate the required scale factors in
TextureProjection::fitTexture(); this turned out to be non-trivial as it would
require a calculation involving the effective aspect ratio of the selected face
in texture space.

Instead, this is done very simply at the GUI level by propagating the
calculated scale factor from the fitted axis to the non-fitted axis, in the
same way a user might achieve this effect manually by copy-pasting scale values
between the respective numeric entry fields.