Use GL_RG8 format for uncompressed normal maps
Now that we are unconditionally reconstructing the Z coordinate in the shader,
there is no need to store it in the texture (and there was never any need for
an alpha channel).

The BindableTexture interface now defines a new Role enum which can be used as
a hint for the implementing image class to choose an appropriate GL internal
format based on usage. In RGBAImage, the role is now used to choose between
GL_RGBA8 for general images (as per the previous behaviour) and GL_RG8 for
normal maps. This appears to save around 500 MB of GPU memory in the St Lucia
map.

Further optimisations would be possible, for example skipping the alpha channel
in diffuse maps if the material does not use any alpha-related commands.