Fix incorrect colours before Colour Schemes dialog shown
The recent switch to threaded loading of entityDefs, while valid in itself,
exposed a bug with entity class colour overrides.

EntityClass::resetColour() did not contain any code to check for colour
overrides in the EClassColourManager, which meant that an overridden colour set
with setColour() could be lost if any subsequent change triggered a call to
resetColour().

The resetColour() method therefore now checks for colour overrides before doing
anything else, and returns immediately if an override colour is set.