Build manual.html in correct output directory
The default behaviour of asciidoctor is to generate the HTML file in the same
directory as the input file, which is incorrect in out-of-source builds. We now
use an explicit -D argument to pass through the CMAKE_CURRENT_BINARY_DIR, so
the HTML file will be generated in the output tree.