Add option for showing offline user guide
Navigating directly to the online website is useful, but having the option to
view the locally-installed version is still valuable for development and
perhaps for some users who cannot or do not wish to rely on online resources.

There are now two entries in the Help menu: "User Guide (Online)" and "User
Guide (Local)" to cover both of these options.