Separate const and modifying Matrix4 column accessors
Distinguish the methods which return a modifiable reference to columns by
appending 'Ref' to the method name, i.e. xCol3Ref(), tColRef() etc. Calling
code is updated to use these versions where a reference is *required*, either
because it is assigned to as an lvalue or returned by const reference to a
higher-level function.

The remaining xCol3(), yCol3() etc methods now return by value. Changing a
method from returning a reference to returning a value is potentially
dangerous, because any calling code which continues to treat the returned value
as if it were a reference (e.g. assigning to it) will do the wrong thing, and
neither G++ nor CLang seem to be able to warn about this. Hopefully the
combination of manual searching along with unit tests has ensured that there
are no remaining uses of xCol3() which should be xCol3Ref().