Streamline Matrix4::translation() and tCol()
translation() was calling tCol().getVector3(), but tCol() was constructing the
Vector4 by appending the tw() component, only to have it stripped off again by
the call to getVector3().

These methods are now flipped around so that translation() uses the Eigen API
to extract the first 3 components directly, and tCol() then adds the tw()
component to create a Vector4.