Fix "linked scaling" button not changing bitmap on GTK
Once again wxGTK seems to be buggy: the "pressed" bitmap on a
wxBitmapToggleButton is only visible when you are actually HOLDING the mouse
button, not when the button is merely toggled. This is now worked around with a
wxutil helper class which explicitly changes the bitmap in response to button
state changes.