Abstract two-column layout construction in SurfaceInspector
Introduce a new class wxutil::FormLayout which wraps the construction of a
sizer containing a two column form-style layout with a column of labels on the
left and an arbitrary widget on the right. This is used in SurfaceInspector
instead of manually constructing wxStaticText labels and packing them into a
wxFlexGridSizer.