Separate SurfaceInspector scale rows from other property rows
FormLayout now uses a wxGridBagSizer internally, which allows a widget to span
more than one row or column. This is now used in a new addFullWidth() method to
add a single widget across both columns. SurfaceInspector makes use of this
method to add a wxStaticLine to separate the horizontal and vertical scale rows
from the previous rows, avoiding the UI weirdness whereby the tightly-linked
scale rows were actually closer to the other manipulator rows than to each
other.