Refactor some EntityTest boilerplate
Add a new helper class TestEntity to construct an entity by classname and add
it to the map (to enable undo). Since this is a struct with public members, we
can use C++17 structured bindings to initialise local variables in tests as
well as accessing members in the struct directly.