Fix building when precompiled headers are turned off.
When building with "cmake -DCMAKE_DISABLE_PRECOMPILE_HEADERS=On",
the build breaks due to several missing includes. This PR fixes that.

(my selfish rationale: precompiled headers seems to break my ccache setup,
and the packaging work for Debian usually includes a tons of rebuilds, so not
be able to use ccache slows me down significantly)