#5836: avoid unwanted colour changes when adjusting multiple lights
This appears to be a long-standing bug with the LightInspector. The
multi-selection code is far too coercive, setting ALL properties from the
dialog on ALL selected lights, rather than just the property which has changed.

This commit fixes the unwanted colour changes when toggling multiple lights
between omni and projected, but the bug still exists in other areas.