#5884: double-clicking a model material no longer accepts dialog
Commit 9cfda496d31a0c5b40 introduced the ability to accept the ModelSelector
and create a model immediately on double-clicking in the model tree, however
the event handler was not bound to the _treeView but to the dialog itself. This
worked, because wxCommandEvents propagate upwards, however it also added the
unwanted behaviour of closing the dialog when double-clicking in the materials
list.

We now Bind() to the _treeView directly rather than to the dialog, ensuring
that only the main model tree can close the dialog with a double-click.