#5885: fix loss of attached entity offset after Reload Defs
To render the entity offset, we are using the localToParent matrix
exposed by the ITransformNode interface. However this matrix was only
being set once, when the attachments were constructed, which isn't
sufficiently "sticky" -- the localToParent matrix can easily be
recalculated by code in the attached entity itself.

We now set the localToParent matrix immediately before rendering each
attached entity, ensuring it is always up to date.