#5901: pointfile identification is now case-insensitive
On Windows there is no guarantee that TDM will write out a .lin file
with the same case as the map name, for example a map named "Leak.map"
may end up with a pointfile "leak_portal_xxx.lin". We now use
case-insensitive string comparisons when identifying matching
pointfiles, so that pointfiles of any case will appear in the list.

This might give wrong results if a mapper on Linux is working on both
"MAP.map" and "map.map" in the same directory, but that seems a very
unlikely situation, and the only ill effect is that the pointfile list
might include more pointfiles than necessary.