Remove Matrix3 column accessor methods
Although reinterpret_cast happens to work with Eigen matrices, it is
inherently dangerous since it relies on assumptions about the internal
layout. These accessor methods were not actually used outside of tests,
and for test purposes they don't offer anything that can't be achieved
using zx(), zy() etc.

The AccessMatrixColumnVectors test is removed entirely (since it only
existed to test these methods), while the call to zCol() in
ConstructTranslationMatrix is removed since the test already compares
the entire list of matrix coefficients with the expected values.