Non-virtual updateGeometry() is now updateGeometryWithData()
The combination of virtual and non-virtual implementing methods is a
solid design, but functions calling overloads of themselves can be
confusing in terms of IDE navigation.

The non-virtual updateGeometry() which accepts vertex and index data is
therefore renamed to updateGeometryWithData() to clearly distinguish it
from the no-argument updateGeometry() which must be overridden by
subclasses.