Simplify RenderVertex constructors
Use a combination of default arguments and templates to reduce the
RenderVertex non-default constructors from 5 to 2. This in turn required
some more flexibility in the BasicVector3/4 constructors: it is now
possible to convert between vectors of different precisions much more
seamlessly.

In theory it ought to be possible to merge the remaining non-templated
RenderVertex constructor with the templated version, but all my attempts
so far resulted in incomprehensible template substitution errors.