#410: add mechanism to grey out invalid menu entries
Connect to the wxEVT_MENU_OPEN event to detect when a menu is about to
be opened, and use the event handler to enable or disable menu items
according to the value returned by ICommandSystem::canExecute().

Currently the only visible change is in the Brush/CSG submenu, since
these are the only items for which canExecute() is implemented.