#410: add support for greying out menu items based on Statements
Some menu items (such as Brush/Prism...) are based on Statements
(commands with arguments) rather than simple Commands. This change
implements the canExecute() method on Statement objects, allowing menu
items such as Prism to be disabled if the underlying command is not
runnable (ignoring arguments).

The only menu items this change currently affects are the first three
items in the Brush menu, which are all variants of the
QueryBrushPrefabSidesDialog command with different arguments.