#6095: add a trivial test for the GameManager
This may not turn out to be all that useful, since the RadiantTest sets
up a fake game by responding on the message bus, and never actually
loads any .game files. Currently it is testing only that the game
manager can return the expected config struct which reflects the fake
game info populated by RadiantTest.