#6095: test accessing some values from .game file
It turns out that the tests ARE loading .game files, although these are
not under the test/resources directory so it remains to be seen whether
this test will encounter any problems on other platforms. Possibly it
will only work if the main DR resources (i.e. install/*) are installed
to their expected location before the tests are run.