#6095: don't create menu items for unsupported game features
The menu.xml allows a new "gamefeature" attribute to be applied to a
<menuItem> node. If present, this feature will be looked up in the
current game using the hasFeature() method, and the menu item will not
be created if the feature is not present.

This is used to hide the "Make Structural" and "Make Detail" menu items
for games which are not Quake 3.