#6120: Add a crashing unit test covering the scenario
After grid-snapping some brushes might be degenerate, and currently these
are not immediately removed the scene. Next time the selection system is
finishing an operation (which in this scenario is a null operation), the
degenerate brushes are removed outside any undoable transaction.
Hitting undo will crash the app due to dangling references.