#6178: ArtProvider is now local to RadiantApp class
LocalBitmapArtProvider was only ever constructed and owned by
RadiantApp; the reason it was exposed as a wxutil header was for the
ArtIdPrefix() method, which was used in a single other header
(Bitmap.h).

LocalBitmapArtProvider is now RadiantApp::ArtProvider, a local nested
class which is not exposed anywhere else. The functions in Bitmap.h just
use the "darkradiant:" prefix explicitly, which could be seen as
duplication but (1) the prefix is very unlikely to change, and (2) the
redundancy isn't serious enough to justify exposing the art provider
class and its method(s) through several headers and their dependent
shared objects.

Also removed the LocalBitmapArtProvider destructor which explicitly
removed the art provider from wxWidgets. This appeared to be redundant
because the wxArtProvider base class already does this in its own
destructor (according to documentation).