#6187: direct wxWidgets log messages to stderr on Linux
The default behaviour of popping up a series of modal dialog boxes is
possibly the most unhelpful way of showing errors imaginable. On Linux
we now use wxLogStderr to direct these log messages to stderr.

On Windows the behaviour is left as-is because stderr isn't usually
accessible; the ideal solution might be to direct the log messages to
the in-application Console window but this would require a custom wxLog
implementation.