#6198: populate ModelSelector after mainframe is constructed
For a reason which isn't clear, the Reparent() call causes problems on
GTK whereby the GUI elements inside the ModelSelector constantly resize
and flicker on mouse motion.

We now avoid the need for a Reparent() altogether by calling Populate()
in response to signal_MainFrameConstructed(), which still allows the
model selector to populate itself but also guarantees that a valid main
frame pointer will be available.