Remove the black "sphere" circle from Rotate manipulator
The black circle represented the ability to perform a trackball-like
arbitrary rotation in screen space. This functionality still exists, but
the black circle itself had no real purpose since you perform the
trackball rotation by clicking on empty space, not on any particular
circle.

The _circleSphere renderable is still present since it is used for mouse
selection, but it no longer renders on screen.