Registry stores keyvalues using XML content (not attributes)
Instead of storing key values in a "value" attribute on a node with no
content, the value is now stored as content and the legacy "value"
attribute (if present) is removed.

To retain backwards compatibility, a "value" attribute will still be
used for retrieval if the node does not contain any content.