#6351: use abort() for unexpected exception on Linux
For a reason which is unclear but probably has something to do with
wxWidgets event handling, the popup-based OnExceptionInMainLoop() is
entering an infinite loop of error popups which cannot be interacted
with or dismissed, which makes debugging impossible on Linux.

Instead we just dump the error message to std::cerr and then abort(),
which creates a console message and then terminates the program. This
code is protected by an __linux__ #ifdef to avoid changing the behaviour
on Windows.