#6332: fix broken OrthoBackgroundPanel spinboxes on Linux
WxWidgets 3.0.5 on GTK seems to default to 0 decimal places for a
wxSpinCtrlDouble, which limits the precision of the sliders (since they
are event-linked to the spinboxes) and makes the +/- buttons useless
since they attempt to increment the spinbox value by less than the
available precision.

This appears to be a WxWidgets bug which does not manifest in the latest
3.2 version, but we can work around it in 3.0 by explicitly setting the
spin controls to have 2 decimal places.