XmlTest no longer depends on RadiantTest on Linux
RadiantTest was only used for a single purpose: locating the test
resources directory. By using TEST_BASE_PATH on Linux, we can avoid the
Catch-22 situation of basic XML tests depending on a module which cannot
be initialised without correctly setting up the XML registry, as well as
improving performance by not initialising the core module for each test.

The original behaviour is retained on Windows because it is not clear
that TEST_BASE_PATH is actually available on that platform (it is
defined in the CMake build system).