Skin Editor decl info line no longer appears on top of Close button
Instead of replacing the wxPanel in the XML form with a DeclFileInfo
panel, we now use the XML form's panel as a container and pack the
DeclFileInfo inside (so it's a wxPanel inside a wxPanel).

It is unclear why the original code was not working correctly, since
wxSizer::Replace() is a supported method and the docs do not suggest
there should be any layout problems after using it (provided you call
Layout(), which we were doing).