ConversationDialog no longer shrinks below minimum panel size
Set the [[nodiscard]] attribute on loadNamedPanel(); this emits a
warning in several dialogs which are ignoring the return value rather
than packing the loaded panel into a sizer. One consequence of this is
that (on GTK at least) the dialog ignores the panel's size hints and can
be shrunk to zero size. This is now corrected for the conversation
editor.