AuiLayout respects windows' minimum sizes
Instead of applying a hard-coded minimum size of 128x128 (which is far
too small for many windows, such as the Surface Inspector), we now ask
the window's main sizer for its minimum size, and set this as the
minimum size of the AUI pane. This makes windows created by the AUI
layout behave more similarly to standard dialogs populated with
SetSizerAndFit().

The possible downside of this change is that if users are docking
widgets like the Surface Inspector (rather than using them as floating
dialogs), they may find that their ability to resize dock panels is more
limited than before, since the docked widget can no longer be shrunk
below the size hint of the widget contents. The Surface Inspector in
particular seems to be claiming a minimum width which is larger than
necessary.