#6514: Skin Name text control only commits changes on ENTER
The previous event was emitted on every character change, which was far
too frequent and led to performance problems and strange behaviour. By
setting the PROCESS_ENTER style flag we can connect instead to an event
which is only emitted when the ENTER key is pressed, which is the more
usual behaviour when editing the name of something in a text field.