Fix RGTC normal maps showing as black in lighting preview mode
Although we were correctly loading RGTC normal maps into the appropriate
destination format, there was no code in the shader to actually handle
the missing Z value, resulting in RGTC normal maps showing as entirely
black.

The Z value is now recovered for all normal maps (because it's much
easier than having to implement format-specific shader behaviour, and
should be fairly quick), using a simple application of Pythagoras. I
can't prove that this is 100% mathematically correct but it seems to
give realistic results, and is a lot better than black.