#6514: add "browse" buttons for Replace/With text boxes
These are using a relative path to the bitmap set directly in the
FBP/XRC file, rather than the "darkradiant:" art provider prefix, which
allows the icons to appear in wxFormBuilder. This works correctly in GTK
since the relative path between XRC and bitmap is the same, and
hopefully will work in Windows too.

The browse buttons show a material chooser to choose a new material for
the associated text box, although the chosen material is not yet being
applied to the skin.