#6514: remove obsolete code from PrefDialog
Remove code relating to expanding nodes and adding sub-pages, since all
pages are now at the same level. Remove function to show Game dialog,
which is never called and would fail since the Game page is no longer
part of this dialog. Switch from wxTreebook to wxListbook, which does
not waste space for non-existent tree expansion controls (although it
does center the page labels, which may not be desirable).